move attach bookmark state func in util
